+++ docx-core/src/reader/namespace.rs
@@ -1,0 +1,28 @@
+// 2006
+pub(crate) const DOC_RELATIONSHIP_TYPE: &str =
+    "http://schemas.openxmlformats.org/officeDocument/2006/relationships/officeDocument";
+pub(crate) const CUSTOM_PROPERTIES_TYPE: &str =
+    "http://schemas.openxmlformats.org/officeDocument/2006/relationships/custom-properties";
+pub(crate) const STYLE_RELATIONSHIP_TYPE: &str =
+    "http://schemas.openxmlformats.org/officeDocument/2006/relationships/styles";
+pub(crate) const NUMBERING_RELATIONSHIP_TYPE: &str =
+    "http://schemas.openxmlformats.org/officeDocument/2006/relationships/numbering";
+pub(crate) const SETTINGS_TYPE: &str =
+    "http://schemas.openxmlformats.org/officeDocument/2006/relationships/settings";
+pub(crate) const COMMENTS_TYPE: &str =
+    "http://schemas.openxmlformats.org/officeDocument/2006/relationships/comments";
+pub(crate) const WEB_SETTINGS_TYPE: &str =
+    "http://schemas.openxmlformats.org/officeDocument/2006/relationships/webSettings";
+pub(crate) const HEADER_TYPE: &str =
+    "http://schemas.openxmlformats.org/officeDocument/2006/relationships/header";
+pub(crate) const FOOTER_TYPE: &str =
+    "http://schemas.openxmlformats.org/officeDocument/2006/relationships/footer";
+pub(crate) const THEME_TYPE: &str =
+    "http://schemas.openxmlformats.org/officeDocument/2006/relationships/theme";
+pub(crate) const IMAGE_TYPE: &str =
+    "http://schemas.openxmlformats.org/officeDocument/2006/relationships/image";
+pub(crate) const HYPERLINK_TYPE: &str =
+    "http://schemas.openxmlformats.org/officeDocument/2006/relationships/hyperlink";
+// 2011
+pub(crate) const COMMENTS_EXTENDED_TYPE: &str =
+    "http://schemas.microsoft.com/office/2011/relationships/commentsExtended";
