+++ docx-core/test_xml_reader.rs
@@ -2,16 +2,19 @@ use std::fs;
 
 // Import from the local crate
 extern crate docx_rs;
-use docx_rs::read_docx_from_xml;
+use docx_rs::read_xml;
 
 fn main() {
     // Read the provided XML file
-    let xml_content = fs::read_to_string("/Users/bokuweb/Downloads/ooxml.xml")
-        .expect("Failed to read XML file");
-    
-    println!("Reading XML file with length: {} characters", xml_content.len());
-    
-    match read_docx_from_xml(&xml_content) {
+    let xml_content =
+        fs::read_to_string("/Users/bokuweb/Downloads/ooxml.xml").expect("Failed to read XML file");
+
+    println!(
+        "Reading XML file with length: {} characters",
+        xml_content.len()
+    );
+
+    match read_xml(&xml_content) {
         Ok(docx) => {
             println!("Successfully parsed XML!");
             println!("Document has {} children", docx.document.children.len());
