+++ docx-wasm/src/reader.rs
@@ -13,7 +13,7 @@ pub fn readDocx(buf: &[u8]) -> Result<String, JsValue> {
 #[allow(non_snake_case)]
 #[wasm_bindgen]
 pub fn readXML(xml: &str) -> Result<String, JsValue> {
-    let mut d = docx_rs::read_docx_from_xml(xml);
+    let mut d = docx_rs::read_xml(xml);
     match d {
         Ok(ref mut d) => Ok(d.json()),
         Err(e) => Err(e.to_string().into()),
